pyproject.toml
[tool.poetry]
name = "nerdlandbot"
version = "0.1.0"
description = ""
authors = ["Arno989", "Autom", "Bob van der Vleuten", "Eva Smeets", "flor01", "Sebastiaan Jansen", "Senne De Vocht", "Thunderb1rdOne", "TsjipTsjip", "Wouter Hermans", "Wouter Meuwis"]

[tool.poetry.dependencies]
python = "^3.8"
discord-py = "^1.5.1"
python-dotenv = "^0.14.0"
pandas = "^1.1.5"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
